Added Unit OF Work Rep

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from 'app/_models/message';
 import { Pagination } from 'app/_models/pagination';
+import { ConfirmService } from 'app/_services/confirm.service';
 import { MessageService } from 'app/_services/message.service';
 
 @Component({
@@ -11,12 +12,12 @@ import { MessageService } from 'app/_services/message.service';
 export class MessagesComponent implements OnInit {
   messages: Message[] = [];
   pagination!: Pagination;
-  container = 'Unread';
+  container = 'Inbox';
   pageNumber = 1;
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -37,9 +38,12 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Confirm Delete Message', 'This Cannot Be Undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        });
+      }
     })
   }
-
 }
